[daemon] Extract settings-file monitoring
To make it available on other platforms (other "daemon mains").
