+++ src/daemon/CMakeLists.txt
@@ -22,6 +22,7 @@ add_library(daemon STATIC
   custom_image_host.cpp
   daemon.cpp
   daemon_config.cpp
+  daemon_monitor_settings.cpp
   daemon_rpc.cpp
   default_vm_image_vault.cpp
   json_writer.cpp
